+++ bundles/catalogue/metadataflyout/view/tab/components/DataQualities.jsx
@@ -2,7 +2,7 @@ import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { Message } from 'oskari-ui';
-import { Label, SubLabel } from '../';
+import { Label, SubLabel } from '..';
 
 const DETAILED = [
     'nameOfMeasure',